// admin/src/app/banner-admin/admin-components/model/product.ts
export class ProductList {
    id?: number;
    product_name?: '';
    category?: '';
    one_unit_weight?: '';
    weight_unit?: '';
    price_type?: '';
    default_product_image?: '';
    is_coupon_allowed?: boolean;
    is_featured?: boolean;
    product_description?: '';
    price_details?: '';
}